+++ server/tests/test_routes_api.py
@@ -78,12 +78,6 @@ class TestLudoboxWebServer(LudoboxTestCase):
         data = json.loads(result.data)
         self.assertEquals(data, ["README.md"])
 
-    def test_game_index(self):
-        result = self.client.get('/api/games')
-        data = json.loads(result.data)
-        self.assertIs(type(data), list)
-        self.assertTrue(type(data[0]), dict)
-
     def test_upload_allowed(self):
         # self.login(email=self.user_email, password=self.user_password)
         self.app.config["UPLOAD_ALLOWED"] = False
@@ -229,58 +223,6 @@ class TestLudoboxWebServer(LudoboxTestCase):
             event = new_game_info["history"][1]
             self.assertEqual(event["type"], "update")
             self.assertEqual(event["user"], self.user_email)
-
-    # sadly, can not test with current user
-    # def test_current_user_profile(self):
-    #     """Check if current user get its profile properly"""
-    #     with self.client:
-    #         self.login()
-    #         result = self.client.get('/api/profile')
-    #         self.assertEqual(result.status_code, 200)
-    #         profile = result.json
-    #         self.assertEqual(profile["email"], self.user_email)
-    #         self.assertEqual(type(profile["recent_changes"]), list)
-    #         self.assertEqual(len(profile["recent_changes"]), 2)
-
-    def test_user_profile(self):
-        """Check if we get profile for any user properly"""
-        with self.client:
-
-            # self.register(email="info@dcalk.org", password="password")
-            # self.login(email="info@dcalk.org")
-
-            result = self.client.get('/api/profile/1')
-            self.assertEqual(result.status_code, 200)
-
-            profile = result.json
-            self.assertEqual(profile["email"], "tester@test.com")
-            self.assertEqual(type(profile["recent_changes"]), list)
-            self.assertEqual(len(profile["recent_changes"]), 1)
-
-    def test_get_recent_changes(self):
-        """Check if we get recent changes"""
-        with self.client:
-
-            result = self.client.get('/api/recent_changes')
-            self.assertEqual(result.status_code, 200)
-            changes = result.json
-            self.assertEqual(len(changes), 2)
-            self.assertEqual(changes[0]["event"]["type"], "create")
-
-            # filter for a single user
-            result = self.client.get('/api/recent_changes?user_id=1')
-            self.assertEqual(result.status_code, 200)
-            changes = result.json
-            self.assertEqual(len(changes), 1)
-            self.assertEqual(changes[0]["event"]["type"], "create")
-
-            # filter limit by time
-            year_2016 = 1451520000 # 31 Dec 2015
-            result = self.client.get('/api/recent_changes?before_time=%s'%year_2016)
-            self.assertEqual(result.status_code, 200)
-            changes = result.json
-            self.assertEqual(len(changes), 1)
-            self.assertEqual(changes[0]["event"]["type"], "create")
 
     # def test_form_add_game(self):
     #     """Posting data and files using form should create a new game"""
